added whatsapp on mobile view for jobs page, gemini function new file
