+++ lib/s3.ts
@@ -1,4 +1,4 @@
-import { S3Client, PutObjectCommand, DeleteObjectCommand } from "@aws-sdk/client-s3";
+import { S3Client, DeleteObjectCommand } from "@aws-sdk/client-s3";
 
 export const s3 = new S3Client({
   region: process.env.AWS_REGION!,
